+++ test/supply_chain.test.js
@@ -93,7 +93,7 @@ contract("SupplyChain", function (accounts) {
           "Struct Item should have a `sku` member"
         );
         assert(
-          isType(subjectStruct)("sku")("uint256"), 
+          isType(subjectStruct)("sku")("uint"), 
           "`sku` should be of type `uint`"
         );
       });
@@ -104,7 +104,7 @@ contract("SupplyChain", function (accounts) {
           "Struct Item should have a `price` member"
         );
         assert(
-          isType(subjectStruct)("price")("uint256"), 
+          isType(subjectStruct)("price")("uint"), 
           "`price` should be of type `uint`"
         );
       });
